[S] Int -> Short

+++ AADWebApp/Interfaces/IBloodTestService.cs
@@ -12,10 +12,10 @@ namespace AADWebApp.Interfaces
         public IEnumerable<BloodTestResult> GetBloodTestResults(short? bloodTestRequestId = null);
         public IEnumerable<BloodTestRequest> GetBloodTestRequests(short? Id = null);
         public IEnumerable<BloodTestRequest> GetBloodTestRequestsByPrescriptionId(short? prescriptionId = null);
-        public Task<int> RequestBloodTestAsync(Prescription prescription, int bloodTestId, DateTime appointmentTime);
-        public Task<int> SetBloodTestDateTimeAsync(Prescription prescription, int bloodTestRequestId, DateTime appointmentTime);
-        public int SetBloodTestResults(int bloodRequestTestId, bool result, DateTime resultTime);
-        public int SetBloodTestRequestStatus(int id, BloodTestRequestStatus bloodTestRequestStatus);
-        public int CancelBloodTestRequest(int id);
+        public Task<int> RequestBloodTestAsync(Prescription prescription, short bloodTestId, DateTime appointmentTime);
+        public Task<int> SetBloodTestDateTimeAsync(Prescription prescription, short bloodTestRequestId, DateTime appointmentTime);
+        public int SetBloodTestResults(short bloodRequestTestId, bool result, DateTime resultTime);
+        public int SetBloodTestRequestStatus(short id, BloodTestRequestStatus bloodTestRequestStatus);
+        public int CancelBloodTestRequest(short id);
     }
 }
