+++ AADWebApp/Models/Prescription.cs
@@ -5,8 +5,8 @@ namespace AADWebApp.Models
 {
     public class Prescription
     {
-        public int Id { get; set; }
-        public int MedicationId { get; set; }
+        public short Id { get; set; }
+        public short MedicationId { get; set; }
         public string PatientId { get; set; }
         public int Dosage { get; set; }
         public DateTime DateStart { get; set; }
